+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -6,7 +6,8 @@ Release History
 * `vpn-connection create/update`: Add `--express-route-gateway-bypass` argument.
 * `express-route`: Port command groups from `express-route` extensions.
 * `express-route`: Added `express-route gateway` and `express-route port` command groups.
-* `express-route create/update`: Added `--express-route-port` argument.
+* `express-route create/update`: Added arguments: `--express-route-port`, `--allow-classic-operation'.
+* `express-route peering create/update`: Added argument `--legacy-mode`.
 
 2.3.2
 +++++